Added logic to fetch a users companies
Along with redirection to create company if no companies exist and redirection to the dashboard if only 1 company exists (no need to select company if only one exists).

Added some default placeholder images to put as company iamges to give some character to the different companies and to make them stand out.